ProductType enum 이동
