+++ VendingMachine/VendingMachine/ProductType.swift
@@ -1,0 +1,19 @@
+//
+//  ProductType.swift
+//  VendingMachine
+//
+//  Created by moon on 2018. 6. 12..
+//  Copyright © 2018년 JK. All rights reserved.
+//
+
+import Foundation
+
+enum ProductType: String {
+    case Coke
+    case Sprite
+    case TOP
+    case Cantata
+    case Georgia
+    case StrawberryMilk
+    case ChocoMilk
+}
